refactor(apps): #510 grid decorator. Delete demo and ast-based version
