--- libs/modules/demo-grid-form/.babelrc
@@ -1,17 +1,0 @@
-{
-  "presets": [
-    "@nrwl/react/babel",
-    "@babel/preset-env",
-    "@babel/preset-react",
-    "@babel/preset-typescript"
-  ],
-  "plugins": [
-    [
-      "@babel/plugin-transform-typescript",
-      {
-        "allowNamespaces": true
-      }
-    ],
-    "@babel/plugin-proposal-class-properties"
-  ]
-}
