--- libs/modules/demo-grid-form/.eslintrc.js
@@ -1,15 +1,0 @@
-module.exports = {
-  extends: '../../../.eslintrc.js',
-  root: true,
-  settings: {
-    'import/resolver': {
-      typescript: {
-        project: 'tsconfig.eslint.json',
-      },
-    },
-  },
-  parserOptions: {
-    tsconfigRootDir: __dirname,
-    project: ['tsconfig.eslint.json'],
-  },
-}
