--- libs/modules/demo-grid-form/README.md
@@ -1,7 +1,0 @@
-# modules-demo-grid-form
-
-This library was generated with [Nx](https://nx.dev).
-
-## Running unit tests
-
-Run `nx test modules-demo-grid-form` to execute the unit tests via [Jest](https://jestjs.io).
