--- libs/modules/demo-grid-form/jest.config.js
@@ -1,12 +1,0 @@
-module.exports = {
-  displayName: 'modules-demo-grid-form',
-  preset: '../../../jest.preset.js',
-  transform: {
-    '^.+\\.[tj]sx?$': [
-      'babel-jest',
-      { cwd: __dirname, configFile: './babel-jest.config.json' },
-    ],
-  },
-  moduleFileExtensions: ['ts', 'tsx', 'js', 'jsx'],
-  coverageDirectory: '../../../coverage/libs/modules/demo-grid-form',
-}
