--- libs/modules/demo-grid-form/src/index.ts
@@ -1,1 +1,0 @@
-export * from './useCases/DemoGridForm'
