--- libs/modules/demo-grid-form/src/useCases/DemoGridForm/DemoGridFormInput.decorators.generated.ts
@@ -1,20 +1,0 @@
-export const DemoGridFormInputDecorators = {
-  email1: {
-    grid: {
-      order: 1,
-      span: 12,
-    },
-  },
-  password0: {
-    grid: {
-      order: 0,
-      span: 16,
-    },
-  },
-  name2: {
-    grid: {
-      order: 2,
-      span: 8,
-    },
-  },
-}
