--- libs/modules/demo-grid-form/src/useCases/DemoGridForm/index.ts
@@ -1,3 +1,0 @@
-export * from './DemoGridFormInput'
-export * from './DemoGridFormInput.generated'
-export * from './DemoGridFormInput.decorators.generated'
