--- libs/modules/demo-grid-form/tsconfig.eslint.json
@@ -1,11 +1,0 @@
-{
-  "extends": "../../../tsconfig.base.json",
-  "include": [
-    "**/*.ts",
-    "**/*.tsx",
-    "**/*.js",
-    "**/*.jsx",
-    "**/*.json",
-    ".*.js"
-  ]
-}
