--- libs/tools/grid-decorator/.storybook/preview.js
@@ -1,5 +1,0 @@
-import { withKnobs } from '@storybook/addon-knobs'
-import { addDecorator } from '@storybook/react'
-import '../../../../.storybook/preview'
-
-addDecorator(withKnobs)
