--- libs/tools/grid-decorator/src/ast-based-decorator/index.ts
@@ -1,2 +1,0 @@
-export * from './gridDecorator'
-export * from './DecoratorMapGenerator'
