+++ libs/tools/grid-decorator/src/lintFiles.ts
