--- libs/tools/grid-decorator/src/reflect-based-decorator/index.ts
@@ -1,1 +1,0 @@
-export * from './gridDecorator'
