--- libs/tools/grid-decorator/src/utils/index.ts
@@ -1,2 +1,0 @@
-export * from './fileMngmtHelpers'
-export * from './lintFiles'
