+++ nx.json
@@ -163,9 +163,6 @@
     "modules-page-stories": {
       "tags": []
     },
-    "modules-demo-grid-form": {
-      "tags": []
-    },
     "tools-grid-decorator": {
       "tags": []
     }
