+++ package.json
@@ -131,7 +131,6 @@
     "shelljs": "0.8.4",
     "supertest": "6.0.1",
     "tsconfig-paths": "^3.9.0",
-    "tslib": "2.0.0",
     "typescript-json-schema": "0.45.1",
     "uuid": "8.3.0",
     "voca": "1.4.0",
