+++ tsconfig.base.json
@@ -110,9 +110,6 @@
       "@codelab/modules/page-stories": [
         "libs/modules/page-stories/src/index.ts"
       ],
-      "@codelab/modules/demo-grid-form": [
-        "libs/modules/demo-grid-form/src/index.ts"
-      ],
       "@codelab/tools/grid-decorator": [
         "libs/tools/grid-decorator/src/index.ts"
       ]
