+++ workspace.json
@@ -1893,29 +1893,6 @@
         }
       }
     },
-    "modules-demo-grid-form": {
-      "root": "libs/modules/demo-grid-form",
-      "sourceRoot": "libs/modules/demo-grid-form/src",
-      "projectType": "library",
-      "architect": {
-        "lint": {
-          "builder": "@nrwl/linter:eslint",
-          "options": {
-            "lintFilePatterns": [
-              "libs/modules/demo-grid-form/**/*.{ts,tsx,js,jsx}"
-            ]
-          }
-        },
-        "test": {
-          "builder": "@nrwl/jest:jest",
-          "outputs": ["coverage/libs/modules/demo-grid-form"],
-          "options": {
-            "jestConfig": "libs/modules/demo-grid-form/jest.config.js",
-            "passWithNoTests": true
-          }
-        }
-      }
-    },
     "tools-grid-decorator": {
       "root": "libs/tools/grid-decorator",
       "sourceRoot": "libs/tools/grid-decorator/src",
@@ -1941,7 +1918,7 @@
           "builder": "@nrwl/workspace:run-commands",
           "options": {
             "commands": [
-              "npx ts-node --project libs/tools/grid-decorator/tsconfig.json -r tsconfig-paths/register libs/tools/grid-decorator/src/reflect-based-decorator/generate.ts"
+              "npx ts-node --project libs/tools/grid-decorator/tsconfig.json -r tsconfig-paths/register libs/tools/grid-decorator/src/generate.ts"
             ],
             "parallel": false
           }
